refactor: split the manifest into seperate files
