+++ Cargo.toml
@@ -50,6 +50,7 @@ regex = "1.10.2"
 reqwest = { version = "0.11.22", default-features = false }
 rip = { package = "rattler_installs_packages", version = "0.1.0", default-features = false }
 serde = "1.0.193"
+serde-untagged = "0.1.5"
 serde_json = "1.0.108"
 serde_spanned = "0.6.4"
 serde_with = { version = "3.4.0", features = ["indexmap"] }
