+++ src/project/manifest/activation.rs
@@ -1,0 +1,7 @@
+use serde::Deserialize;
+
+#[derive(Default, Clone, Deserialize, Debug)]
+#[serde(deny_unknown_fields)]
+pub struct Activation {
+    pub scripts: Option<Vec<String>>,
+}
