+++ src/project/manifest/metadata.rs
@@ -1,0 +1,54 @@
+use crate::utils::spanned::PixiSpanned;
+use rattler_conda_types::{Channel, Platform, Version};
+use serde::Deserialize;
+use serde_with::{serde_as, DisplayFromStr};
+use std::path::PathBuf;
+use url::Url;
+
+/// Describes the contents of the `[package]` section of the project manifest.
+#[serde_as]
+#[derive(Debug, Clone, Deserialize)]
+#[serde(deny_unknown_fields)]
+pub struct ProjectMetadata {
+    /// The name of the project
+    pub name: String,
+
+    /// The version of the project
+    #[serde_as(as = "Option<DisplayFromStr>")]
+    pub version: Option<Version>,
+
+    /// An optional project description
+    pub description: Option<String>,
+
+    /// Optional authors
+    #[serde(default)]
+    pub authors: Vec<String>,
+
+    /// The channels used by the project
+    #[serde_as(deserialize_as = "Vec<crate::project::serde::ChannelStr>")]
+    pub channels: Vec<Channel>,
+
+    /// The platforms this project supports
+    // TODO: This is actually slightly different from the rattler_conda_types::Platform because it
+    //     should not include noarch.
+    pub platforms: PixiSpanned<Vec<Platform>>,
+
+    /// The license as a valid SPDX string (e.g. MIT AND Apache-2.0)
+    pub license: Option<String>,
+
+    /// The license file (relative to the project root)
+    #[serde(rename = "license-file")]
+    pub license_file: Option<PathBuf>,
+
+    /// Path to the README file of the project (relative to the project root)
+    pub readme: Option<PathBuf>,
+
+    /// URL of the project homepage
+    pub homepage: Option<Url>,
+
+    /// URL of the project source repository
+    pub repository: Option<Url>,
+
+    /// URL of the project documentation
+    pub documentation: Option<Url>,
+}
